Add proposal test builder
