+++ shared_model/builders/protobuf/proposal.hpp
@@ -18,80 +18,8 @@
 #ifndef IROHA_PROTO_PROPOSAL_BUILDER_HPP
 #define IROHA_PROTO_PROPOSAL_BUILDER_HPP
 
-#include "backend/protobuf/queries/proto_query.hpp"
-#include "builders/protobuf/unsigned_proto.hpp"
-#include "interfaces/common_objects/types.hpp"
-#include "interfaces/iroha_internal/proposal.hpp"
-#include "interfaces/transaction.hpp"
+#include "builders/protobuf/builder_templates/proposal_template.hpp"
 
-#include "ordering.pb.h"
+using ProposalBuilder = TemplateProposalBuilder<>;
 
-namespace shared_model {
-  namespace proto {
-    template <int S = 0>
-    class TemplateProposalBuilder {
-     private:
-      template <class T>
-      using w = detail::PolymorphicWrapper<T>;
-      using TransactionContainer = std::vector<w<Transaction>>;
-
-      template <int>
-      friend class TemplateProposalBuilder;
-
-      enum RequiredFields { Transactions, Height, CreatedTime, TOTAL };
-
-      template <int s>
-      using NextBuilder = TemplateProposalBuilder<S | (1 << s)>;
-
-      iroha::ordering::proto::Proposal proposal_;
-
-      template <int Sp>
-      TemplateProposalBuilder(const TemplateProposalBuilder<Sp> &o)
-          : proposal_(o.proposal_) {}
-
-      /**
-       * Make transformation on copied content
-       * @tparam Transformation - callable type for changing the copy
-       * @param t - transform function for proto object
-       * @return new builder with updated state
-       */
-      template <int Fields, typename Transformation>
-      auto transform(Transformation t) const {
-        NextBuilder<Fields> copy = *this;
-        t(copy.proposal_);
-        return copy;
-      }
-
-     public:
-      TemplateProposalBuilder() = default;
-
-      auto height(const interface::types::HeightType height) const {
-        return transform<Height>(
-            [&](auto &proposal) { proposal.set_height(height); });
-      }
-
-      auto transactions(const TransactionContainer &transactions) const {
-        return transform<Transactions>([&](auto &proposal) {
-          for (const auto &tx : transactions) {
-            proposal.mutable_transactions()->Add(tx->getTransport());
-          }
-        });
-      }
-
-      auto createdTime(
-          const interface::types::TimestampType created_time) const {
-        return transform<CreatedTime>(
-            [&](auto &proposal) { proposal.set_created_time(created_time); });
-      }
-
-      Proposal build() {
-        // TODO 22/01/2018 x3medima17: add stateless validator IR-836
-        static_assert(S == (1 << TOTAL) - 1, "Required fields are not set");
-        return Proposal(iroha::ordering::proto::Proposal(proposal_));
-      }
-    };
-
-    using ProposalBuilder = TemplateProposalBuilder<>;
-  }  // namespace proto
-}  // namespace shared_model
-#endif  // IROHA_PROTO_PROPOSAL_BUILDER_HPP
+#endif //IROHA_PROTO_PROPOSAL_BUILDER_HPP
